xero/requirements.txt
discord.py==1.5
websockets==8.1
aiohttp==3.6.2
PyNaCl==1.3.0
multidict==4.7.5
async-timeout==3.0.1
pymongo==3.10.1
dnspython==1.16.0
aiohttp==3.6.2
asyncio
pytz
timeago==1.0.13
requests
discord_webhook
colorama
DiscordUtils
yarl
chardet
libneko
psutil
